+++ components/experimental/Cargo.toml
@@ -54,7 +54,6 @@ num-bigint = { workspace = true }
 num-rational = { workspace = true, features = ["num-bigint"]}
 num-traits = { workspace = true }
 serde = { workspace = true, features = ["derive", "alloc"], optional = true }
-serde_borrow_utils = { workspace = true, optional = true }
 smallvec = { workspace = true, features = ["const_generics"] }
 
 icu_experimental_data = { workspace = true, optional = true }
@@ -75,7 +74,7 @@ default = ["compiled_data"]
 compiled_data = ["dep:icu_experimental_data", "icu_decimal/compiled_data", "icu_list/compiled_data", "icu_plurals/compiled_data", "icu_properties/compiled_data", "icu_normalizer/compiled_data", "icu_casemap/compiled_data", "icu_provider/baked"]
 datagen = ["serde", "dep:databake", "zerovec/databake", "zerotrie/databake", "tinystr/databake", "icu_collections/databake", "log", "icu_pattern/databake", "icu_plurals/datagen", "icu_pattern/alloc", "icu_provider/export"]
 ryu = ["fixed_decimal/ryu"]
-serde = ["dep:serde", "dep:serde_borrow_utils", "zerovec/serde", "potential_utf/serde", "tinystr/serde", "icu_collections/serde", "icu_decimal/serde", "icu_list/serde", "icu_pattern/serde", "icu_plurals/serde", "icu_provider/serde", "zerotrie/serde", "icu_normalizer/serde", "icu_casemap/serde"]
+serde = ["dep:serde", "zerovec/serde", "potential_utf/serde", "tinystr/serde", "icu_collections/serde", "icu_decimal/serde", "icu_list/serde", "icu_pattern/serde", "icu_plurals/serde", "icu_provider/serde", "zerotrie/serde", "icu_normalizer/serde", "icu_casemap/serde"]
 
 [[bench]]
 name = "transliterate"
