--- utils/serde_borrow_utils/Cargo.toml
@@ -1,22 +1,0 @@
-# This file is part of ICU4X. For terms of use, please see the file
-# called LICENSE at the top level of the ICU4X source tree
-# (online at: https://github.com/unicode-org/icu4x/blob/main/LICENSE ).
-
-[package]
-name = "serde_borrow_utils"
-version = "0.0.0"
-
-rust-version.workspace = true
-authors.workspace = true
-edition.workspace = true
-repository.workspace = true
-homepage.workspace = true
-license.workspace = true
-categories.workspace = true
-include.workspace = true
-
-[dependencies]
-serde = { workspace = true }
-
-[dev-dependencies]
-serde_json = { workspace = true }
