--- utils/serde_borrow_utils/README.md
@@ -1,11 +1,0 @@
-# serde_borrow_utils [![crates.io](https://img.shields.io/crates/v/serde_borrow_utils)](https://crates.io/crates/serde_borrow_utils)
-
-<!-- cargo-rdme start -->
-
-`serde_borrow_utils` is a utility crate of the [`ICU4X`] project.
-
-<!-- cargo-rdme end -->
-
-## More Information
-
-For more information on development, authorship, contributing etc. please visit [`ICU4X home page`](https://github.com/unicode-org/icu4x).
